setup.py: Include `aflak' executable in distribution
The program can now be run simply by typing `aflak' in your terminal
after having been installed with setuptools.

It seems that setuptools require a function as an entry point. That's
why I forcefully added a function to __main__.py.
The function does nothing and the script will simply be run almost as if
`python aflak' had been run on the `aflak' package.