Extend pyqtgraph's classes: Use log-scale on histogram LUT
The only effective difference in behavior with this commit is as
follows:
- Histogram LUT's y coordinates are shown using a logarithmic scale

In astronomical images, the difference between the lowest and highest
luminosities recorded per image is astronomical. Hence the need to use a
log scale.

From a programming perspective, this commit is a breakthrough as it
allows to some extent to copy-pasta-less code re-use of pyqtgraph,
while enabling customization.
Care was taken to extend each of the pyqtgraph classes of which we are
changing the behavior.

AstroImageViewTemplate.ui is written from the ImageViewTemplate.ui file
found in pyqtgraph. We are just adapting the template to our needs to
replacing `HistogramLUTWidget' with `AstroHistogramLUTWidget'.
From there, AstroHistogramLUTWidget uses AstroHistogramLUTItem, which
implements a histogram using a log-scale, and here we have the result.