mainwindow_ui: Work around missing text feature for separators
Though text for separators can be set up from QtDesigner, the result
seem to be ignored and not included into the .ui file.

I tried fiddling with the UI file myself, but pyuic would raise compile
errors or plainly ignore the text string.

It is terrible, but add the missing text during Python code generation
in the Makefile.