Makefile: General .deb files
Using the command below, binary .deb files for supported platforms will
be generated:

  make all-deb

make-deb.sh: Build debian binary packages for several debian derivatives

The major difference between distro seems to be the Python version that
is required. Depending on the version of Python, the install folder is
indeed not the same.

Currently only tested on ubuntu17.10.