AstroImageView: Work around bug in range on showing arrows
When arrows are shown in the canvas (via pg.ArrowItem), the default range
of the axis becomes larger. The reasons behind this strange behaviour are
unknown.

Calling auto-range again after automatic ranging fixes the issue, but is
a dirty work around that may cause other unexpected behaviors in the
future.