WCSAxis: Fix axis generation with no FITS file
As axis is generated through information in the FITS header, a FITS file
is needed to generate the axes in world coordinations.

This commit check if FITS file is loaded before generating axes.
Indeed, the program would crash when no FITS file was loaded on axis
generation. When no FITS file is loaded, axis generation use the
built-in behaviour from *pg.AxisItem*.